+++ source/bronze/src/bronze/application/config/spark_session.py
@@ -1,0 +1,7 @@
+from pyspark import SparkConf
+from pyspark.sql.session import SparkSession
+
+
+def initialize_spark() -> SparkSession:
+    spark_conf = SparkConf(loadDefaults=True).set("spark.sql.session.timeZone", "UTC")
+    return SparkSession.builder.config(conf=spark_conf).getOrCreate()
