+++ source/bronze/src/bronze/application/streams/submitted_transactions.py
@@ -1,21 +1,9 @@
-from pyspark import SparkConf
-from pyspark.sql.session import SparkSession
-
+import bronze.application.config.spark_session as spark_session
+import bronze.infrastructure.streams.kafka_stream as kafka_stream
 from bronze.application.settings.submitted_transactions_stream_settings import SubmittedTransactionsStreamSettings
-from bronze.domain.constants.database_names import DatabaseNames
-from bronze.domain.constants.table_names import TableNames
 
 
 def submit_transactions() -> None:
+    spark = spark_session.initialize_spark()
     kafka_options = SubmittedTransactionsStreamSettings().create_kafka_options()
-
-    spark = initialize_spark()
-
-    spark.readStream.format("kafka").options(**kafka_options).load().writeTo(
-        f"{DatabaseNames.bronze_database}.{TableNames.bronze_submitted_transactions_table}"
-    )
-
-
-def initialize_spark() -> SparkSession:
-    spark_conf = SparkConf(loadDefaults=True).set("spark.sql.session.timeZone", "UTC")
-    return SparkSession.builder.config(conf=spark_conf).getOrCreate()
+    kafka_stream.submit_transactions(spark, kafka_options)
